requirements.txt
python-dotenv
pygame
pylint